v2.51 Release Attempt #1

+++ src/libsodium/__main__.py
@@ -93,7 +93,7 @@ S:::::::::::::::SS  oo:::::::::::oo   d:::::::::ddd::::di::::::i  uu::::::::uu::
  SSSSSSSSSSSSSSS      ooooooooooo      ddddddddd   dddddiiiiiiii    uuuuuuuu  uuuummmmmm   mmmmmm   mmmmmm
 """
     print("\x1b[32m"+x+"\x1b[0m")
-    print("v2.50\nMade by ahsan")
+    print("v2.51\nMade by ahsan")
     exit()
 
 if args[0] == "init":
@@ -184,6 +184,11 @@ if __name__ == "__main__":
     from concurrent import futures
     addAll(grpc.server(futures.ThreadPoolExecutor(max_workers=10)))""")
 
+    os.mkdir("src/templates")
+    x = open(prefix+"src/templates/.sodium", "w")
+    x.write(os.getcwd()+f"/{project_name}")
+    x.close()
+
     x = open(prefix+"start.py", 'w')
     x.write('''from libsodium import Deamon
 from sonora.wsgi import grpcWSGI
@@ -251,7 +256,7 @@ S:::::::::::::::SS  oo:::::::::::oo   d:::::::::ddd::::di::::::i  uu::::::::uu::
 
     """
     print(F_Green+x+F_End)
-    print("v2.50")
+    print("v2.51")
     print(f"{getCurrentTime()} [{F_Magenta}INFO{F_End}] Creating Deamon... ")
     MainDeamon = Deamon()
 
@@ -590,7 +595,7 @@ class {answer1}JwtVerifier:
             print(f"The utility {args[2]} is not creatable")
             exit()
     elif args[1] == "gRPC":
-        if not len(args) == 3:
+        if not len(args) >= 3:
             print("Invalid amount of arugments please check out the code below\npython3 -m libsodium create gRPC example.proto")
             exit(1)
         try:
@@ -627,11 +632,14 @@ class {answer1}JwtVerifier:
             pass
         
         exit_code = os.system(f"{interpreter} -m grpc_tools.protoc -I src/gRPC/protobufs --python_out=src/gRPC/{args[2][:len(args[2])-6]} --grpc_python_out=src/gRPC/{args[2][:len(args[2])-6]} src/gRPC/protobufs/{args[2]}")
-
         if not exit_code == 0:
             print("Process is shutting down")
-            os.rmdir(f"src/gRPC/{args[2][:len(args[2])-6]}")
+            if not args[3] == "--regen":
+                os.rmdir(f"src/gRPC/{args[2][:len(args[2])-6]}")
             exit(1)
+
+        if args[3] == "--regen":
+            exit()
 
         f = open(f"src/gRPC/{args[2][:len(args[2])-6]}/{args[2][:len(args[2])-6]}.py", "w")
         f.write(f"""import src.gRPC.{args[2][:len(args[2])-6]}.{args[2][:len(args[2])-6]}_pb2_grpc as {args[2][:len(args[2])-6]}_pb2_grpc
